IT WORKS
both RX are still fighting to get into the same staging array. To combat this we dedicate one channel to identify which TX is the data coming from, and storing them on their own arrays before delivering them on rcChannels.data

The caveat is that we are definitely losing some refresh rates of the rc channels, further testing is needed to determine how failsafe works.